+++ src/configure/karma/index.js
@@ -6,6 +6,7 @@ export default function configureKarma (config) {
 
     frameworks: ['jasmine', 'sinon'],
     browsers: ['PhantomJS'],
+    reporters: ['mocha'],
 
     files: [
       'src/**/*.spec.*',
@@ -17,7 +18,10 @@ export default function configureKarma (config) {
       'src/**/*.spec.*': ['webpack']
     },
 
-    webpack: webpackConfig,
+    // there can be multiple archetypes configured
+    // and althought harmless to have them all running the tests
+    // it is not required and only produces double execution
+    webpack: webpackConfig[0],
 
     webpackServer: {
       noInfo: true