+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+### Master
+
+There were two small breaking changes, making this upgrade very simple:
+
+- The plugin `webpack-pages` was renamed to `webpack-archetype-pages`;
+- We changed the way webpack configurations are merged, now loaders are merged toghether if their test name matches.
+
+Changes:
+
+- [#36](https://github.com/pirelenito/sagui/pull/36) Add support to Library projects;
+- [#35](https://github.com/pirelenito/sagui/pull/35) Use webpack-merge smart feature to allow modifying an existing loader.
+
+### v3.3.0 (2016-03-10)
+
+- [#51](https://github.com/pirelenito/sagui/pull/51) Nicer test reporter with "karma mocha reporter"
+
+### v3.2.0 (2016-03-03)
+
+- [#47](https://github.com/pirelenito/sagui/pull/47) Bootstrap gitignore file
+- [#49](https://github.com/pirelenito/sagui/pull/49) Resolve .jsx modules by default
+
 ### v3.0.0 (2016-02-21)
 
 Allow extensibility and customization of webpack and karma configurations.